* amd/info_file.c: add key to sun_entry2amd()'s parameters
* amd/info_hesiod.c: add key to sun_entry2amd()'s parameters

* amd/info_ldap.c: add key to sun_entry2amd()'s parameters

* amd/info_ndbm.c: add key to sun_entry2amd()'s parameters

* amd/info_nis.c: add key to sun_entry2amd()'s parameters

* amd/info_nisplus.c: add key to sun_entry2amd()'s parameters

* amd/sun_map.c: New functions to support sun translation to amd.
The function sun_entry2amd's parameters where expanded to include
the key value.  sun_entry_2amd() now returns NULL on error.

* amd/sun_map.h: Expand sun_entry2amd's parameters to include key
value.

* amd/sun_map_parse.y: Change how file system types are parsed out
of the list of options for each entry.  The previous way was not
working properly so now we just look for the 'fstype' keyword in
each option.  If it exists we copy its key value.