Currently, the mount options (struct am_opts) are associated with the mntfs structure. This is usually not a problem since these options are not used once the mount has been completed. However, if two lookups are within a short time it's possible that the options derived during the second lookup will overwrite the option information in the mntfs structure and thus the first lookup's options are lost.
We've seen plenty of problems where two keys which referenced the same
share but had different sublinks ended up pointing to the same
subdirectory of the share.

This patch introduces a new structure linking an mntfs to a node and
moves the am_opts pointer to this structure.

This is what we currently have:

    +---------+ 1   n +-------+
    | am_node |-------| mntfs |
    +---------+       +-------+
                          |
                     +---------+
                     | am_opts |
                     +---------+

This patch changes the structure above into something like this:

    +---------+ 1   n +--------+ n   1 +-------+
    | am_node |-------| am_loc |-------| mntfs |
    +---------+       +--------+       +-------+
                          |                |
                      +---------+     +---------+
                      | am_opts |     | am_opts |
                      +---------+     +---------+

In this schema am_opts is a per-node structure so that parallel
lookups don't overwrite each other's options. The am_opts structure
associated with mntfs is filled in when the actual mount happens.

(Note that this structure is far from being perfect. The problem is
that am_opts contains two types of fields: fields which are
mount-specific (rhost, rfs, etc.) and fields which would rather belong
to the node (sublink). The diagram above means that we're
unnecessarily duplicating all mount-specific information for each
node.)

From: Krisztian Kovacs <Kris.Kovacs@morganstanley.com>