Make sure to remove nodes in the proper order when going down.
Depending on what order the nodes got created it's possible that
the parent of a node has a bigger am_mapno (index in exported_ap[])
so that it gets freed before its child while the child's am_parent
pointer is still pointing to the already freed block of memory.

This change makes sure that umount_exported() cleans up all children
of a node first before freeing the node.

XXX: This patch did not apply cleanly because it pre-dated the am_al
patches.

From: Krisztian Kovacs <Kris.Kovacs@morganstanley.com>