From: Ian Kent <ikent@redhat.com>
For some reason, when umounting autofs mounts after closing the ioctl
file handle, the kernel can return EBUSY for some small amount of time.

This can cause umounts to incorrectly fail when in fact they should
succeed.

Retrying for about a second and a half seems to work quite well.